use doubles for tie points

+++ s2tbx-rapideye-reader/src/main/java/org/esa/beam/dataio/rapideye/RapidEyeL1Reader.java
@@ -199,7 +199,7 @@ public class RapidEyeL1Reader extends RapidEyeReader {
         readerMap.put(band, reader);
     }
 
-    private TiePointGrid addTiePointGrid(int width, int height, Product product, String gridName, float[] tiePoints) {
+    private TiePointGrid addTiePointGrid(int width, int height, Product product, String gridName, double[] tiePoints) {
         final TiePointGrid tiePointGrid = createTiePointGrid(gridName, 2, 2, 0, 0, width, height, tiePoints);
         product.addTiePointGrid(tiePointGrid);
         return tiePointGrid;
